Move benchmark tests to benchmarks folder
No behavior change, you can still manually test with
`python benchmarks/benchmarks/canonical.py -v`

- Move more common code to common.py
- Migrate to snake_case
- Make DATASETS_DIR less hardcoded
- Ensure settings files are also in datasets,
  before they varied depending on CWD
- Reformat with black
- Apply flake8 to benchmarks too
